Use full grid paths in gbasf2 output queries to avoid gbasf2 bug (#41)
Explanation: The command `gb2_ds_get` is used to download the output data from
the grid and `gb2_ds_list` can be used to list directory contents on the grid.
Both take grid paths and allow for wildcards and are mostly symmetrical in their
usage. So after downloading some data from the grid, I have a check that uses
`gb2_ds_list` to check the remote data on the grid and see if it is the same as
the downloaded directory contents.

If given a relative directory path, both commands automatically check for that
directory in /belle/user/<username>. However, in the current gbasf2 version,
`gb2_ds_list` returns garbage to stdout when you use wildcards in relative
paths. But with absolute paths it seems to work, so I added the
`/belle/user/<username>` prefix to the query string again.

Co-authored-by: Michael Eliachevitch <meliache@uni-bonn.de>